Deleted redundant Bar component. Moved export to utils
